✨ Add Product module (CRUD + filtering UI)
- Implement Product entity mapped to products table
- Add ProductRepository with QueryDSL-based queries
- Create ProductService for clean business logic separation
- Introduce ProductController + ProductView (JavaFX)
- Filtering by Segment (Residential/Business) and Category (TV, Net, Phone, etc.)
- Integrated with main navigation (HelloController + Config)
- Updated persistence.xml to include Product entity
- Connected to main SQLite database (data/db/flamingo.db)

+++ src/main/java/hu/flamingo/app/model/User.java
@@ -9,7 +9,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     @Column(name = "last_name", nullable = false)
     private String lastName;
@@ -24,7 +24,7 @@ public class User {
     public User() {
     }
 
-    public User(int userId, String lastName, String firstName, boolean isActive) {
+    public User(Long userId, String lastName, String firstName, boolean isActive) {
         this.userId = userId;
         this.lastName = lastName;
         this.firstName = firstName;
@@ -38,11 +38,11 @@ public class User {
     }
 
     // ----- Getters & Setters -----
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
