+++ src/main/java/hu/flamingo/app/repository/interfaces/IGenericRepository.java
@@ -7,11 +7,11 @@ public interface IGenericRepository<T> {
 
     void save(T entity);
 
-    Optional<T> findById(int id);
+    Optional<T> findById(Long id);
 
     List<T> findAll();
 
     void update(T entity);
 
-    void delete(int id);
+    void delete(Long id);
 }
